--- common/list_pdf_utils/utils/__init__.py
@@ -1,1 +1,0 @@
-from .file_downloaders import downloaders
